+++ CHANGELOG.rst
@@ -1,8 +1,11 @@
-Version next
+Version 0.3.0
   * Fixed bug due to changes in the OSX backend for matplotlib 2.0.
   * Addressed deprecation warnings due to changes in matplotlib 2.0 in a
     backwards compatible way with matplotlib 1.5.
   * Updated color scheme to align with new matplotlib 2.0 defaults.
+  * Fixed bug that can mess up axis limits when one column doesn't have data
+  * Changed some calls to axis plot methods to explicitly call the axis method instead of relying on pyplot to get it right.
+  * Add a hack to fix the scaling of the lower-right 1D panel for matplotlib versions < 2.0.
 
 Version 0.2.4
   * Fixed bug that was causing a crash in Python 3.5